Improved Enemy
Now there is an enemy prefab, and multiple enemies can be spawned and chase the player simultaneously.
NOTE:
Whenever you change the platform/level/terrain, be sure to click the Level gameobject containing all the terrain and on the left going into navigation and then clicking bake. This updates the navmesh for the enemies to be able to pathfind.